make deposit gen frequency configurable and add preset datapacks to control it, add locate all command, fix config reload issue, fix drill rendering in contraption

+++ src/main/java/com/bmaster/createrns/content/deposit/mining/MiningBehaviour.java
@@ -234,5 +234,5 @@ public abstract class MiningBehaviour extends BlockEntityBehaviour implements ID
         return true;
     }
 
-    public record MinerSpec(ClaimingArea miningArea, float miningSpeed) {}
+    public record MinerSpec(ClaimingArea miningArea, double miningSpeed) {}
 }
